// app/providers/mocks.d.ts
export declare const MOCKS: {
    "tickets/counts": {
        "new_messages": number;
        "open_all": number;
        "open_as_tech": number;
        "open_as_alttech": number;
        "open_as_user": number;
        "onhold": number;
        "reminder": number;
        "parts_on_order": number;
        "unconfirmed": number;
        "waiting": number;
    };
    "todos": {
        "list_id": string;
        "id": any;
        "name": string;
        "text": string;
        "title": string;
        "order_list": number;
        "order_item": number;
        "assigned_id": any;
        "assigned_name": string;
        "type": number;
        "ticket_id": number;
        "project_id": number;
        "account_id": number;
        "due_date": any;
        "updated_at": any;
        "is_completed": boolean;
        "estimated_remain": any;
        "ticket_number": any;
        "time_hours": any;
        "time_is_billable": boolean;
        "time_task_type_id": any;
        "time_task_type_name": string;
        "time_payment_id": any;
        "time_invoice_id": any;
        "time_start_time": any;
        "time_stop_time": any;
        "sub": {
            "list_id": string;
            "id": string;
            "name": string;
            "text": string;
            "title": string;
            "order_list": number;
            "order_item": number;
            "assigned_id": number;
            "assigned_name": string;
            "type": number;
            "ticket_id": number;
            "project_id": number;
            "account_id": number;
            "due_date": string;
            "updated_at": string;
            "is_completed": boolean;
            "estimated_remain": number;
            "ticket_number": any;
            "time_hours": any;
            "time_is_billable": boolean;
            "time_task_type_id": any;
            "time_task_type_name": string;
            "time_payment_id": any;
            "time_invoice_id": any;
            "time_start_time": any;
            "time_stop_time": any;
            "sub": any;
        }[];
    }[];
    "login": {
        "api_token": string;
    };
    "organizations": {
        "key": string;
        "name": string;
        "is_expired": boolean;
        "is_trial": boolean;
        "instances": {
            "key": string;
            "name": string;
            "is_expired": boolean;
            "is_trial": boolean;
        }[];
    }[];
    "config": {
        stat: {};
        "key": string;
        "org": string;
        "instance": string;
        "is_onhold_status": boolean;
        "is_time_tracking": boolean;
        "is_freshbooks": boolean;
        "freshbooks_url": string;
        "is_parts_tracking": boolean;
        "is_project_tracking": boolean;
        "is_unassigned_queue": boolean;
        "is_location_tracking": boolean;
        "is_waiting_on_response": boolean;
        "is_invoice": boolean;
        "is_payments": boolean;
        "is_expenses": boolean;
        "is_class_tracking": boolean;
        "is_travel_costs": boolean;
        "is_priorities_general": boolean;
        "is_confirmation_tracking": boolean;
        "is_resolution_tracking": boolean;
        "is_ticket_levels": boolean;
        "is_account_manager": boolean;
        "is_require_ticket_initial_post": boolean;
        "is_ticket_require_closure_note": boolean;
        "is_asset_tracking": boolean;
        "assets": {
            "unique1_caption": string;
            "unique2_caption": string;
            "unique3_caption": string;
            "unique4_caption": string;
            "unique5_caption": string;
            "unique6_caption": string;
            "unique7_caption": string;
        };
        "time_hour_increment": number;
        "time_minimum_time": number;
        "timezone_offset": number;
        "timezone_name": string;
        "currency": string;
        "businessday_length": number;
        "logo": string;
        "user": {
            "account_name": string;
            "login_id": string;
            "user_id": number;
            "email": string;
            "firstname": string;
            "lastname": string;
            "is_techoradmin": boolean;
            "is_admin": boolean;
            "is_limit_assigned_tkts": boolean;
            "is_useworkdaystimer": boolean;
            "account_id": number;
            "time_format": number;
            "date_format": number;
        };
    };
};